% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{box}
\alias{box}
\title{Make a content box for a shiny app}
\usage{
box(title = "", ..., icon = NULL)
}
\arguments{
\item{title}{Title of the box}

\item{icon}{Icon to be placed to the left of title (must use e.g. icon("home"))}
}
\description{
Make a content box for a shiny app
}
