% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{dashboard_body}
\alias{dashboard_body}
\title{Make a body for a Shinto app bs4Dash::dashboardBody with various common dependencies (shinyjs, shinytoastr,
bootstrap icons, etc.)}
\usage{
dashboard_body(
  ...,
  busybar_color = "#53C082",
  loadingscreen_time = 2,
 
    disconnect_message = "Oeps, er ging iets fout! Herlaad de pagina en probeer het nog een keer."
)
}
\arguments{
\item{\dots}{UI arguments passed to bs4Dash::dashboardBody}

\item{busybar_color}{Color of 'busy bar'}

\item{loadingscreen_time}{Loading time (not yet implemented)}

\item{disconnect_message}{Message for disconnect popup}
}
\description{
Wrapper around
}
