% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{dashboard_header}
\alias{dashboard_header}
\title{Make the header UI for a Shinto app}
\usage{
dashboard_header(
  title = "Shinto app framework",
  icon = shiny::icon("home"),
  ...
)
}
\arguments{
\item{title}{Title of the application (top-left)}

\item{icon}{Icon next to title}

\item{\dots}{Further UI elements to be placed in the right part of the header}
}
\description{
Make the header UI for a Shinto app
}
