% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{dashboard_sidebar}
\alias{dashboard_sidebar}
\title{Make a sidebar for a Shinto app}
\usage{
dashboard_sidebar(...)
}
\arguments{
\item{\dots}{All arguments passed to bs4Dash::sidebarMenu}
}
\description{
Make a sidebar for a Shinto app
}
