% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatafel.R
\name{datatafel}
\alias{datatafel}
\title{Datatable, Nederlandse versie}
\usage{
datatafel(
  ...,
  dom = "frtip",
  pageLength = 10,
  escape = FALSE,
  rownames = FALSE,
  ordering = TRUE,
  extensions = "Responsive",
  selection = "single",
  scrollX = FALSE,
  options = list()
)
}
\description{
Voor shintoui, gebruikt style='bootstrap4'
}
