% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{menu_item}
\alias{menu_item}
\title{Make a Menu item for a Shinto app}
\usage{
menu_item(text, icon, tabName, ...)
}
\arguments{
\item{text}{Text to be displayed in the menu}

\item{icon}{Icon for menu item (must use e.g. icon("home"))}

\item{tabName}{Name of the tab, must refer to a tabitem.}
}
\description{
Make a Menu item for a Shinto app
}
