% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBoxModule.R
\name{plotBoxUI}
\alias{plotBoxUI}
\alias{plotBox}
\title{Shiny module for a plot and the data in a tab box}
\usage{
plotBoxUI(id, title = "", ...)

plotBox(
  input,
  output,
  session,
  data = reactive(NULL),
  plot_function = function(data) NULL,
  table_function = function(data) data,
  digits = 0,
  align = "l",
  striped = TRUE
)
}
\description{
Shiny module for a plot and the data in a tab box
}
