% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shintoui_dependencies.R
\name{shintoui_dependencies}
\alias{shintoui_dependencies}
\title{Browser/navigator info dependencies}
\usage{
shintoui_dependencies()
}
\description{
In the UI, place the command \code{shintoui::shintoui_dependencies()}.
}
