% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{tab_panel}
\alias{tab_panel}
\alias{tabset_panel}
\alias{tab_items}
\alias{tab_item}
\title{Make a tabset panel (Bootstrap 4)}
\usage{
tab_panel(...)

tabset_panel(...)

tab_items(...)

tab_item(tabName, ...)
}
\description{
Make a tabset panel (Bootstrap 4)
}
